backend/requirements.txt
flask==3.0.0
werkzeug==3.0.0
flask-cors==4.0.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
certifi>=2024.2.2
urllib3>=2.0.0
pyOpenSSL>=23.0.0
requests>=2.31.0